XmPrivate.h: never try to build this automatically.
XmPrivate must be generated manually.  For this you work, you must
have a freshly compiled openmotif tree, and MLIBSRC must be pointing
to it.

Otherwise, it's possible during the includes phase for an attempt to
be made to regenerate this file, which will fail on the vast majority
of systems out there.

So, to regenerate,

cd include/Xm
rm XmPrivate.h
make XmPrivate.h