Revert "Enable the use of the TIRPC library on Linux."
This reverts commit 8a8619bfa8a25342d898d2cc4decd7bb66b9c060.

More work will need to be done to use tirpc on 64bit systems.  It
works 'accidentally' on 32b systems.  The issue is that tirpc
includes *must* be used, and there is some work required to properly
support this (like proper include paths, proper definition of XDR,
etc).

So for now, we revert this until that work can be completed and tested,
otherwise 64b linux builds are likely to have problems.