Fix dtcm build on OpenBSD-current.
libcompat and header files will be gone in the upcoming 5.5 release, so we
cannot use the ftime(3) interface there.  Fall back to the SVR4 version of the
code, but use the tm_gmtoff field of struct tm to get the timezone.
Unfortunately, that is not portable (because I really really would like to
remove that old and crummy struct timeb from the public API).