Remove old cruft: osf/* and utils/prog*
The osf/ stuff includes uil, wml, etc - we expect these to be provided
by Motif, and make no attempt to build them in CDE, so no need for
them.

The only program in utils/prog/ is an ancient verion of lndir that we
do not need for CDE.  If you need lndir for some reason, use the
laters version from Xorg, usually installed via the xutils-dev
package.