dtfile/File.c: implement a hack to fix Ticket #19, tree icons are black
The real issue seems to be a bug in Motif.  The background color for
these tree icons is always black.  Depending on what Palette you have
selected, it's possible for the foreground color to be black.  When
this happens, you will see a black square since both fg and bg are now
black.

You can select another Palette that works (ie: foreground is white)
and the problem goes away.  So, for now, we always force a white
foreground color so the actual symbols are visible in tree mode.