FreeBSD 11: Don't use imported x11/motif, use installed versions for build
With this patch, the import symlinks for x11 and motif do not need to
be created.  The build will use the installed headers in
/usr/local/include/{X11, Xm}.

This will work for FreeBSD 11 only ATM as I don't have older systems
to test with.

It may work for FreeBSD 10 or earlier versions as well.  If so, edit
config/cf/FreeBSD.cf and change the OSMajorVersion check for
UseInstalledX11 appropriately.  And send a patch to the CDE mailing
list :)