OpenBSD 6+: use installed X11 and Motif
This commit removes the need to setup X11 and motif import symlinks
before building CDE.  With OpenBSD v6+, the installed versions of X11 and
Motif will be used.

This may work with earlier versions of OpenBSD as well, but I don't
have access to those older versions.  If you try it, and it works,
send a patch to the CDE mailing list.

This was tested on OpenBSD 6.2