Add a new script, dtapp, used to locate and run various helper programs
This script is located in /usr/dt/bin/dtapp, and is then symlinked to
various helpers that can be used in DT actions to run programs.

The various helpers currently installed are:

dtapp_vimage - view an image file, override with DTAPP_VIMAGE.
Defaults to xv, display, and gimp, in that order.

dtapp_vpdf - view a PDF file, override with DTAPP_VPDF.  Defaults to
okular, xpdf

dtapp_vps - view a postscript file, override with DTAPP_VPS.  Defaults
to mgv, gv

dtapp_vvideo - view a video file, override with DTAPP_VVIDEO.
Defaults to vlc, ffplay.

When a request is made to view one of these files, the list of viewers
will be tried, in order, until one is found.  If none are found, an
error message will be displayed.  Add overrides to your ~/.dtprofile
file.

We can add more dtapp commands and defaults for them as needed.

This is in preparation for integrating Antonis Tsolomitis' extended
actions and icon files, coming up in future commits.