Clean up OPENBUGS
This patch fixes some wierd word splicing that occured with the importing of
the sources, which caused the last letter of many bugs to be cut off, with
some of them cutting off more, I made my best judgement on what was supposed
to be there.

This patch also gets rid of old bugs and software that CDE no longer includes.